+++ packages/spacecat-shared-utils/src/auth.js
@@ -12,6 +12,8 @@
 
 import AWSXray from 'aws-xray-sdk';
 import { GetSecretValueCommand, SecretsManagerClient } from '@aws-sdk/client-secrets-manager';
+import { Site } from '@adobe/spacecat-shared-data-access';
+import { ImsPromiseClient } from '@adobe/spacecat-shared-ims-client';
 import { isString } from './functions.js';
 import { resolveCustomerSecretsName } from './helpers.js';
 
@@ -20,14 +22,43 @@ import { resolveCustomerSecretsName } from './helpers.js';
  */
 
 /**
+ * Get an access token by exchanging a promise token using IMS Promise Client.
+ * @param {object} context - The context object containing environment variables.
+ * @param {string} promiseToken - The promise token to exchange for an access token.
+ * @return {Promise<object>} - A promise that resolves to the access token response.
+ */
+export async function getAccessToken(context, promiseToken) {
+  const imsClient = ImsPromiseClient.createFrom(
+    context,
+    ImsPromiseClient.CLIENT_TYPE.CONSUMER,
+  );
+
+  const token = await imsClient.exchangeToken(
+    promiseToken,
+    !!context.env?.AUTOFIX_CRYPT_SECRET && !!context.env?.AUTOFIX_CRYPT_SALT,
+  );
+
+  return token.access_token;
+}
+
+/**
  * Retrieves the page authentication token for a given site.
  *
  * @param {Site} site - The site to retrieve authentication for
  * @param {object} context - The context object
- * @returns {Promise<string>} - The authentication token needed to access the page
+ * @param {object} authOptions - The authentication options
+ * @returns {Promise<string>} - The authentication token or access token
  * @throws {Error} - If secret is not found or token is missing
  */
-export async function retrievePageAuthentication(site, context) {
+export async function retrievePageAuthentication(site, context, authOptions = {}) {
+  const CS_TYPES = [Site.AUTHORING_TYPES.CS, Site.AUTHORING_TYPES.CS_CW];
+  if (site
+    && (CS_TYPES.includes(site.getAuthoringType())
+      || site.getDeliveryType() === Site.DELIVERY_TYPES.AEM_CS)
+    && authOptions.promiseToken) {
+    return getAccessToken(context, authOptions.promiseToken.promise_token);
+  }
+
   const baseURL = site.getBaseURL();
   const customerSecret = resolveCustomerSecretsName(baseURL, context);
   const secretsManagerClient = new SecretsManagerClient({});