+++ packages/spacecat-shared-utils/src/index.d.ts
@@ -43,6 +43,8 @@ export function toBoolean(value: unknown): boolean;
 
 export function isValidUrl(urlString: string): boolean;
 
+export function isValidHelixPreviewUrl(urlString: string): boolean;
+
 export function isValidUUID(uuid: string): boolean;
 
 export function isValidIMSOrgId(imsOrgId: string): boolean;
@@ -154,6 +156,17 @@ declare function generateCSVFile(data: object[]): Buffer;
  * @returns {string} - The content with placeholders replaced.
  */
 declare function replacePlaceholders(content: string, placeholders: object): string;
+
+/**
+ * Function to support reading static file
+ * and replace placeholder strings with values.
+ *
+ * @param {Object} placeholders - A JSON object containing values to replace in the prompt content.
+ * @param {String} filename - The path of the prompt file.
+ * @returns {Promise<string|null>} - A promise that resolves to a string with the prompt content.
+ */
+declare function getStaticContent(placeholders: object, filename: string):
+    Promise<string | null>;
 
 /**
  * Reads the content of a prompt file asynchronously and replaces any placeholders